Refactor reminders to use class-based commands and PDO
Replaces old procedural reminder commands with new class-based Reminder and ReminderList commands using PDO for database access. Moves Dota constants and reminder utilities to BotUtils.php, updates Main.php to use the new reminder check, and removes legacy code from OldCommands. This improves maintainability and consistency across the codebase.
